+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- constructing a searchspace containing a `TaskParameterwith` the 
+  `SearchSpace.from_dataframe` constructor method now raises an `IncompatibilityError`
 - `NumericalTarget.match_*` constructors now accept a `mismatch_instead` argument. If 
   set to `True`, targets seek to avoid the given `match_value` instead of matching it.
 - `NumericalTarget.match_*` constructors now accept a `match_mode` argument. While `"="`